pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-repair-estimator"
version = "0.1.0"
description = "VK bot and backend for auto body repair cost estimation with optional ML pipeline."
requires-python = ">=3.11"
authors = [
  { name = "AutoRepairEstimator Developer" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "vkbottle>=4.3.0",
  "aiokafka>=0.10.0",
  "asyncpg>=0.29.0",
  "minio>=7.2.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.2.0",
  "numpy>=1.26.0",
  "Pillow>=10.0.0",
  "opencv-python-headless>=4.8.0",
  "ultralytics>=8.0.0",
  "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "anyio[trio]>=4.0.0",
  "pytest-cov>=4.1.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
  "mypy>=1.9.0",
  "types-Pillow>=10.0.0",
  "testcontainers[postgres]>=4.8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "W",
  "F",
  "I",
  "B",
  "C4",
  "UP",
  "SIM",
]
ignore = ["E501", "B008", "UP042"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.coverage.run]
omit = [
    "*/bot/main.py",
    "*/bot/notification_consumer.py",
    "*/bot/backend_client.py",
    "*/bot/config.py",
    "*/bot/handlers/*.py",
    "*/bot/keyboards/*.py",
    "*/ml_worker/main.py",
    "*/adapters/db/connection.py",
    "*/adapters/gateways/kafka_producer.py",
    "*/adapters/gateways/kafka_consumer.py",
    "*/adapters/gateways/minio_storage_gateway.py",
    "*/workers/outbox_flusher.py",
    "*/workers/heartbeat_checker.py",
    "*/workers/inference_result_consumer.py",
    "*/ml_worker/inference/result_publisher.py",
    "*/ml_worker/inference/damage_detector.py",
    "*/ml_worker/inference/parts_detector.py",
    "*/ml_worker/s3_client.py",
    "*/ml_worker/config.py",
    "*/ml/train_*.py",
    "*/ml/evaluate.py",
    "*/ml/export_model.py",
]

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = [
  "-v",
  "--strict-markers",
]
markers = [
  "anyio: mark test as an anyio async test",
  "integration: tests that require external infrastructure (Docker / Postgres container)",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
  "asyncpg.*",
  "aiokafka.*",
  "minio.*",
  "ultralytics.*",
  "cv2.*",
  "vkbottle.*",
  "vkbottle_types.*",
]
ignore_missing_imports = true

